requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
daphne==4.1.0
distro==1.9.0
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-anymail==13.0
django-redis==5.4.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
kombu==5.5.2
msgpack==1.0.8
openai==1.74.0
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
